+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For a full diff see [`2.3.2...main`][2.3.2...main].
 
 - Extracted `Duration` ([#351]), by [@localheinz]
 - Merged `MaximumDuration` into `Duration` ([#352]), by [@localheinz]
+- Renamed `MaximumCount` to `Count` ([#353]), by [@localheinz]
 
 ### Fixed
 
@@ -176,5 +177,6 @@ For a full diff see [`7afa59c...1.0.0`][7afa59c...1.0.0].
 [#350]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/350
 [#351]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/351
 [#352]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/352
+[#353]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/353
 
 [@localheinz]: https://github.com/localheinz